removed redundant conversion

+++ x/interchainstaking/keeper/proposal_handler.go
@@ -12,7 +12,6 @@ import (
 	icatypes "github.com/cosmos/ibc-go/v5/modules/apps/27-interchain-accounts/types"
 	ibcexported "github.com/cosmos/ibc-go/v5/modules/core/exported"
 	tmclienttypes "github.com/cosmos/ibc-go/v5/modules/light-clients/07-tendermint/types"
-
 	"github.com/ingenuity-build/quicksilver/x/interchainstaking/types"
 )
 
@@ -63,7 +62,7 @@ func (k *Keeper) HandleRegisterZoneProposal(ctx sdk.Context, p *types.RegisterZo
 		DepositsEnabled:    p.DepositsEnabled,
 		Decimals:           p.Decimals,
 		UnbondingPeriod:    int64(tmClientState.UnbondingPeriod),
-		MessagesPerTx:      int64(p.MessagesPerTx),
+		MessagesPerTx:      p.MessagesPerTx,
 	}
 	k.SetZone(ctx, zone)
